// part_9/patientor_backend/types.ts
export interface Diagnose {
  code: string;
  name: string;
  latin?: string;
}

export interface Patient {
  id: string;
  name: string;
  dateOfBirth: string;
  ssn: string;
  gender: string;
  occupation: string;
}

export type NewPatientEntry = Omit<Patient, 'id'>;

export type NonSensitivePatientEntry = Omit<Patient, 'ssn'>;